backend/requirements.slim.txt
# Core dependencies (minimal install for running the project)
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.1.1
pydantic==2.11.9
starlette==0.37.2
python-multipart==0.0.20
emergentintegrations==0.1.0
google-genai==1.42.0
